Add files via upload
This is a project I have been working on in which I scrape data primarily from the world bank, but others as well, and use the data to analyze GDP growth.  The python notebook contains the majority of the code I used for this project, and I explain my analysis in the pdf file.